+++ projects/rocprofiler-compute/src/argparser.py
@@ -147,6 +147,28 @@ Examples:
         "--target", type=str, default=None, help=argparse.SUPPRESS
     )
     profile_group.add_argument(
+        "--attach-pid",
+        type=str,
+        dest="attach_pid",
+        metavar="",
+        default=None,
+        required=False,
+        help="\t\t\tProcess id to be attached for profiling.",
+    )
+    profile_group.add_argument(
+        "--attach-duration-msec",
+        type=str,
+        dest="attach_duration_msec",
+        metavar="",
+        default=None,
+        required=False,
+        help=(
+            "\t\t\tWhen --attach-pid is used, it specifies the attach duration "
+            "in milliseconds. If not set, detachment occurs when "
+            '"Enter" key is pressed.'
+        ),
+    )
+    profile_group.add_argument(
         "-p",
         "--path",
         metavar="",